STRAT-3377 — Bump the motd
Based on feedback from $CEO, Stratus is having its name shifted
(slightly) to become "stratus maas".

This commit reflects that change in the motd that is used when users
login to the terminal.

== Design Notes
=== Untested

The author is unclear exactly how this repository works, but has
submitted a patch in what they believe to be the correct way.